Hotfix for v1.0 -> v.1.1 compatibility
If you had v1.0 installed while trying to install v1.1, the installation would fail.
This hotfix fixed that issue.

The problem was, that v1.0 used the backup file ending ".ori" while v.1.1 uses ".ucp_backup". Thus, the new setup was not able to find the original Crusader executable, if v1.0 was still installed.